+++ ouroboros-network-testing/src/Ouroboros/Network/Testing/Data/Script.hs
@@ -16,6 +16,7 @@ module Ouroboros.Network.Testing.Data.Script
   , initScript'
   , stepScript'
   , stepScriptSTM'
+  , shrinkScriptWith
   , arbitraryScriptOf
   , prop_shrink_Script
     -- * Timed scripts
@@ -112,16 +113,18 @@ stepScriptSTM' scriptVar = do
 
 instance Arbitrary a => Arbitrary (Script a) where
     arbitrary = sized $ \sz -> arbitraryScriptOf sz arbitrary
+    shrink = shrinkScriptWith shrink
 
-    shrink (Script (x :| [])) = [ Script (x' :| []) | x' <- shrink x ]
-    shrink (Script (x :| xs)) =
+shrinkScriptWith :: (a -> [a]) -> Script a -> [Script a]
+shrinkScriptWith f (Script (x :| [])) = [ Script (x' :| []) | x' <- f x ]
+shrinkScriptWith f (Script (x :| xs)) =
         Script (x :| [])                          -- drop whole tail
       : Script (x :| take (length xs `div` 2) xs) -- drop half the tail
       : Script (x :| init xs)                     -- drop only last
 
         -- drop none, shrink only elements
-      : [ Script (x' :| xs) | x'  <- shrink x ]
-     ++ [ Script (x :| xs') | xs' <- shrinkVector shrink xs ]
+      : [ Script (x' :| xs) | x'  <- f x ]
+     ++ [ Script (x :| xs') | xs' <- shrinkVector f xs ]
 
 
 --